Relocate `ApplicationTemplateCommand` `CreateAppCommand` from `org.grails.cli.commands` to `org.grails.cli.command.app`
