+++ grace-shell/src/main/groovy/org/grails/cli/command/app/ApplicationTemplateCommand.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.cli.commands
+package org.grails.cli.command.app
 
 import groovy.transform.CompileDynamic
 import groovy.transform.CompileStatic
@@ -33,6 +33,11 @@ import org.grails.build.logging.GrailsConsoleAntBuilder
 import org.grails.build.logging.GrailsConsoleAntProject
 import org.grails.build.logging.GrailsConsoleLogger
 import org.grails.build.parsing.CommandLine
+import org.grails.cli.commands.CommandDescription
+import org.grails.cli.commands.ExecutionContext
+import org.grails.cli.commands.ProjectCommand
+import org.grails.cli.commands.ProjectContext
+import org.grails.cli.commands.ProjectContextAware
 import org.grails.cli.profile.ProfileRepository
 import org.grails.cli.profile.ProfileRepositoryAware
 import org.grails.cli.profile.repository.MavenProfileRepository
